5. Подсчет функции в диапазоне. [#24251 #248505]

+++ src/main/java/ru/job4j/lambda/FunctionDiapason.java
@@ -7,8 +7,8 @@ import java.util.function.Function;
 public class FunctionDiapason {
     public static List<Double> diapason(int start, int end, Function<Double, Double> func) {
         List<Double> rsl = new ArrayList<>();
-        for (double i = start; i < end; i++) {
-            rsl.add(func.apply(i));
+        for (int i = start; i < end; i++) {
+            rsl.add(func.apply((double) i));
         }
         return rsl;
     }
